refactor(twofactor): refine finalize challenge function

+++ internal/twofactor/twofactor_service.go
@@ -63,12 +63,6 @@ func (s *TwoFactorService) getUserState(ctx context.Context, stateID string) (*U
 	return userState, err
 }
 
-type ChallengeOptions struct {
-	Subject     Subject
-	RedirectURL string
-	ExpiresIn   time.Duration
-}
-
 func (s *TwoFactorService) CreateChallenge(ctx context.Context, subject Subject, callbackURL string, expiresIn time.Duration) (*Challenge, error) {
 	ch := Challenge{
 		ID:          uuid.NewString(),
@@ -123,59 +117,69 @@ func (s *TwoFactorService) ValidateChallenge(ctx context.Context, ch *Challenge,
 
 type verifyFunc func(userState *UserState) (bool, error)
 
-func (s *TwoFactorService) verifyChallenge(ctx context.Context, ch *Challenge, sub Subject, doChallengerVerify verifyFunc) error {
+func (s *TwoFactorService) verifyChallenge(ctx context.Context, ch *Challenge, sub Subject, doChallengerVerify verifyFunc) (string, error) {
 	stateID := s.calculateHash(sub.UserID, sub.IPAddress)
 	userState, err := s.getUserState(ctx, stateID)
 	if err != nil {
-		return err
+		return "", err
 	}
 
 	userState.FailCount, err = s.userStateStore.IncreaseFailCount(ctx, stateID)
 	if err != nil {
-		return err
+		return "", err
 	}
 	if userState.FailCount > params.TwoFactorMaxFailCount {
-		return ErrTooManyFailedAttempts
+		return "", ErrTooManyFailedAttempts
 	}
 
 	ch.Attempts, err = s.challengeStore.IncreaseAttempts(ctx, ch.ID)
 	if err != nil {
-		return err
+		return "", err
 	}
 	if ch.Attempts > params.TwoFactorChallengeMaxAttempts {
-		return ErrChallengeAttemptsExceeded
+		return "", ErrChallengeAttemptsExceeded
 	}
 
 	success, err := doChallengerVerify(userState)
 	if err != nil {
-		return err
+		return "", err
 	}
 	if success {
+
 		if err := s.challengeStore.MarkSuccess(ctx, ch.ID); err != nil {
-			return err
+			return "", err
+		}
+
+		verifiedAt := time.Now()
+		if err := s.challengeStore.SetVerifiedAt(ctx, ch.ID, verifiedAt); err != nil {
+			return "", err
 		}
 		s.userStateStore.ResetFailCount(ctx, stateID)
 		s.userStateStore.DecreaseChallengeCount(ctx, stateID)
-		return nil
+		token := s.calculateHash(ch.ID, verifiedAt.UnixNano())
+		return token, nil
 	}
 
 	attemptsLeft := min(params.TwoFactorChallengeMaxAttempts-ch.Attempts, params.TwoFactorMaxFailCount-userState.FailCount)
 	if attemptsLeft == 0 {
-		return ErrTooManyFailedAttempts
+		return "", ErrTooManyFailedAttempts
 	}
-	return NewAttemptFailError(attemptsLeft)
+	return "", NewAttemptFailError(attemptsLeft)
 }
 
-func (s *TwoFactorService) FinalizeChallenge(ctx context.Context, cid string, sub Subject) error {
+func (s *TwoFactorService) FinalizeChallenge(ctx context.Context, cid string, sub Subject, callbackURL string) error {
 	ch, err := s.challengeStore.Get(ctx, cid)
 	if errors.Is(err, store.ErrNotFound) {
 		return ErrChallengeNotFound
 	}
+	if ch.Success == 0 {
+		return ErrChallengeNotVerified
+	}
 	if ch.Subject != s.subjectHash(sub) {
 		return ErrChallengeSubjectMismatch
 	}
-	if ch.Success == 0 {
-		return ErrChallengeNotVerified
+	if ch.CallbackURL != callbackURL {
+		return ErrChallengeCallbackURLMismatch
 	}
 	err = s.challengeStore.Delete(ctx, ch.ID)
 	if errors.Is(err, store.ErrNotFound) {
